# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "HufflePuff's NodeBucket - OpenAPI Documentation",
    "description": "This is a Task Management Server documented according to the OpenAPI Specification.\n\nSome useful links:\n- [NodeBucket repository](https://github.com/just-enginerin/nodebucket)"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "tags": [
    {
      "name": "Employees",
      "description": "All operations available regarding employees."
    }
  ],
  "paths": {
    "/employees/{empId}": {
      "get": {
        "tags": ["Employees"],
        "summary": "findEmployeeById",
        "description": "Get an employee by their ID.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "A numerical value between 1007 - 1012",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "empId": {
                      "type": "number",
                      "example": 1007
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Johann"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Bach"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/employees/{empId}/tasks": {
      "get": {
        "tags": ["Employees"],
        "summary": "findAllTasks",
        "description": "Get all tasks created for a designated employee.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "A numerical value between 1007 - 1012",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "empId": {
                      "type": "number",
                      "example": 1007
                    },
                    "todo": {
                      "type": "array",
                      "example": []
                    },
                    "done": {
                      "type": "array",
                      "example": []
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Employees"],
        "summary": "createTask",
        "description": "Create a new task for a designated employee.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "A numerical value between 1007 - 1012",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Information about the new task.",
          "required": true,
          "content": {
          "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "type": "object",
                    "properties": {
                      "text": {
                        "type": "string",
                        "required": true,
                        "description": "A brief title describing the new task.",
                        "example": "Test Task"
                      },
                      "category": {
                        "type": "object",
                        "required": true,
                        "description": "The designated category of the new task.",
                        "properties": {
                          "categoryName": {
                            "type": "string",
                            "required": true,
                            "description": "The name of the new task's designated category.",
                            "example": "meetings"
                          },
                          "backgroundColor": {
                            "type": "string",
                            "required": true,
                            "description": "The label color associated with the new task's designated category.",
                            "example": "FFFFFF"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "64e93dfb9f0e3c3b7fa5bc0d"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["Employees"],
        "summary": "updateTask",
        "description": "Update the task list for a designated employee.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "A numerical value between 1007 - 1012",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Information about the task list.",
          "required": true,
          "content": {
          "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "todo": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "required": true,
                          "description": "The unique ID of the task.",
                          "example": "64e6bb3d7ff5447cb9892ccc"
                        },
                        "text": {
                          "type": "string",
                          "required": true,
                          "description": "A brief title describing the task.",
                          "example": "To do Test Task"
                        },
                        "category": {
                          "type": "object",
                          "required": true,
                          "description": "The designated category of the task.",
                          "properties": {
                            "categoryName": {
                              "type": "string",
                              "required": true,
                              "description": "The name of the task's designated category.",
                              "example": "meetings"
                            },
                            "backgroundColor": {
                              "type": "string",
                              "required": true,
                              "description": "The label color associated with the task's designated category.",
                              "example": "FFFFFF"
                            }
                          }
                        }
                      }
                    }

                  },
                  "done": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "required": true,
                          "description": "The unique ID of the task.",
                          "example": "64e6b2477ff5447cb9892cc9"
                        },
                        "text": {
                          "type": "string",
                          "required": true,
                          "description": "A brief title describing the task.",
                          "example": "Done Test Task"
                        },
                        "category": {
                          "type": "object",
                          "required": true,
                          "description": "The designated category of the task.",
                          "properties": {
                            "categoryName": {
                              "type": "string",
                              "required": true,
                              "description": "The name of the task's designated category.",
                              "example": "projects"
                            },
                            "backgroundColor": {
                              "type": "string",
                              "required": true,
                              "description": "The label color associated with the task's designated category.",
                              "example": "FFFFFF"
                            }
                          }
                        }
                      }
                    }

                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success; No content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/employees/{empId}/tasks/{taskId}": {
      "delete": {
        "tags": ["Employees"],
        "summary": "deleteTasks",
        "description": "Delete a task.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "A numerical value between 1007 - 1012",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "description": "A unique ID for the designated task.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success; No content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"]
}